Chat.tsxのヘルパー関数を分離
