+++ app/client/src/utils/chat.ts
@@ -1,0 +1,38 @@
+// チャットで利用する共通ユーティリティ
+
+export function isUrl(value?: string): boolean {
+  if (!value) return false;
+  try {
+    const url = new URL(value.trim());
+    return url.protocol === "http:" || url.protocol === "https:";
+  } catch {
+    return false;
+  }
+}
+
+export type ActorID = string;
+
+export function splitActor(actor: ActorID): [string, string | undefined] {
+  if (actor.startsWith("http")) {
+    const url = new URL(actor);
+    return [url.pathname.split("/").pop()!, url.hostname];
+  }
+  if (actor.includes("@")) {
+    const [user, domain] = actor.split("@");
+    return [user, domain];
+  }
+  return [actor, undefined];
+}
+
+export function normalizeActor(actor: ActorID): string {
+  if (actor.startsWith("http")) {
+    try {
+      const url = new URL(actor);
+      const name = url.pathname.split("/").pop()!;
+      return `${name}@${url.hostname}`;
+    } catch {
+      return actor;
+    }
+  }
+  return actor;
+}
